requirements.txt
streamlit
torch
whisper
langchain_groq
langchain_core
python-dotenv
langchain_community
ffmpeg-python